Define DossierJS on `window`.
The previous approach of using `var` *should* work (since the `var` will
get hoisted to the enclosing scope), but this will fail in contexts
where the enclosing scope isn't global.

Well, in that case, `var` would be working as intended, right? Yes, but:

1) We don't always control that context.
2) Despite that, we still really need the module to be available in the
   global scope.

If you want to do it "right," then use AMD. Otherwise, we seem to be
forced to play it fast & loose in the global scope.

This *could* be a breaking change if you're embedding DossierJS in
a non-global scope where there is some existing but different global
variable named `DossierJS`. One can only hope that you do not find
yourself in this precarious situation!

[breaking-change]